pyproject.toml
[project]
name = "writing-style-analyzer"
version = "0.1.0"
description = "A local writing style analyzer using LLMs for German and English text"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "pyyaml>=6.0",
    "langdetect>=1.0.9",
    "tqdm>=4.66.0",
    "sentencepiece>=0.2.0",
    "accelerate>=0.30.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.1.0",
    "odfpy>=1.4.1",
    "pylatexenc>=2.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=7.0.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
]
llama-cpp = [
    "llama-cpp-python>=0.2.0",
]
# Optional LLM validation testing dependencies
# Install only if you want to run LLM validation tests
llm-validation = [
    "anthropic>=0.39.0",
    "openai>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["analyze.py", "config.yaml"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pyyaml>=6.0.3",
    "ruff>=0.14.2",
]
